+++ packages/error/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @u-moni/error
 
+## 2.1.4
+
+### Patch Changes
+
+- add SDK_NAME and SDK_VERSION
+- Updated dependencies
+  - @u-moni/common@1.1.4
+
 ## 2.1.3
 
 ### Patch Changes